backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation and serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Pin to 4.0.1 for passlib compatibility

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
nhl-api-py==3.1.1

# Yahoo Fantasy API
yahoo-oauth
yahoo-fantasy-api

# Background tasks (optional, for future)
# celery==5.3.4
# redis==5.0.1
